requirements.txt
aiogram
pandas
openpyxl
sqlalchemy
psycopg2-binary
Flask
python-dotenv